requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
mysql-connector-python
python-dotenv
EOL
